Move all factories to a single file, update tests.
