+++ p3/tests/test_cart.py
@@ -4,8 +4,7 @@ from django.core.urlresolvers import reverse
 from django.test import TestCase
 from django_factory_boy import auth as auth_factories
 
-from conference.tests.factories.attendee_profile import AttendeeProfileFactory
-from conference.tests.factories.conference import ConferenceFactory
+from tests.factories import AttendeeProfileFactory, ConferenceFactory
 
 
 class TestCartView(TestCase):
