+++ tests/test_admin.py
@@ -6,8 +6,7 @@ from django.core import mail
 from django.core.urlresolvers import reverse
 from django_factory_boy import auth as auth_factories
 
-from assopy.stripe.tests import factories
-from conference.tests.factories.conference import ConferenceFactory
+from tests import factories
 
 
 @mark.django_db
@@ -68,7 +67,7 @@ def test_order_invoices_admin(admin_client):
 
 @mark.django_db
 def test_order_stats_admin(admin_client):
-    ConferenceFactory()
+    factories.ConferenceFactory()
     url = reverse('admin:assopy-order-stats')
 
     response = admin_client.get(url)
