+++ tests/test_assopy_templates_paths.py
@@ -22,10 +22,10 @@ from django_factory_boy import auth as auth_factories
 from pytest import mark
 
 from assopy.models import Invoice, Order, Vat
-from assopy.tests.factories.user import AssopyUserFactory
 from conference.models import AttendeeProfile
 
 from tests.common_tools import template_paths
+from tests.factories import AssopyUserFactory
 
 
 def make_sure_root_template_is_used(response, template_name):
