+++ tests/test_exchangerates.py
@@ -1,7 +1,3 @@
-# coding: utf-8
-
-
-
 from datetime import date
 from decimal import Decimal
